@@ -44,15 +44,15 @@
     "lodash": "^4.17.20",
     "nanoevents": "^5.1.8",
     "prop-types": "^15.7.2",
-    "react": "^16.12.0",
+    "react": "^17.0.2",
     "react-click-outside": "^3.0.1",
     "react-color": "^2.18.0",
-    "react-dom": "^16.12.0",
+    "react-dom": "^17.0.2",
     "react-dropzone": "^11.0.1",
     "react-final-form": "^6.3.5",
     "react-hotkeys": "^2.0.0",
     "react-redux": "^7.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.0.2",
     "react-snap": "^1.23.0",
     "redux": "^4.0.5",
     "redux-actions": "^2.6.5",
@@ -88,7 +88,7 @@
     "less-loader": "^7.0.2",
     "prettier": "2.5.0",
     "react-addons-test-utils": "^15.6.2",
-    "react-test-renderer": "^16.12.0",
+    "react-test-renderer": "^17.0.2",
     "redux-mock-store": "^1.5.4",
     "style-loader": "^3.3.1",
     "ts-jest": "^26.4.1",
