@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import * as canvasApi from '../srcCanvas/api';
 import { updateCanvasSize } from './model/canvas/canvasActions';
 import * as doc from './services/document';
@@ -18,14 +18,10 @@ canvasApi.imageUpdated.on(onImageUpdatedCb);
 export const App = () => (
   <Provider store={store}>
     <Router>
-      <Switch>
-        <Route exact path="/">
-          <AppView />
-        </Route>
-        <Route exact path="/about">
-          <AboutView />
-        </Route>
-      </Switch>
+      <Routes>
+        <Route path="/" element={ <AppView /> } />
+        <Route path="/about" element={<AboutView />} />
+      </Routes>
     </Router>
   </Provider>
 );
